requirements.txt
cffi==1.17.0
clr-loader==0.2.6
configparser==7.1.0
enums==0.0.2
psycopg2==2.9.9
pycparser==2.22
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
pythonnet==3.0.3
shiboken6==6.7.2
